requirements.txt
kafka-python
Faker
mysql-connector-python==8.0.28
#requests
